1.上云部署配置更新

+++ kieslect-common/kieslect-common-security/src/main/java/com/kieslect/common/security/service/TokenService.java
@@ -126,7 +126,7 @@ public class TokenService
 
     public void delLoginUserByUserkey(String userkey)
     {
-        if (StringUtils.isNotEmpty(userkey))
+        if (StringUtils.isNotEmpty(userkey) && redisService.hasKey(getTokenKey(userkey)))
         {
             redisService.deleteObject(getTokenKey(userkey));
         }
